requirements.txt
Flask~=3.0.2
flask-cors~=4.0.1
Flask-SQLAlchemy
flask-swagger
flask-swagger-ui
psycopg2~=2.9.9
SQLAlchemy
python-dotenv
pandas
